+++ database/migrations/2024_08_27_031053_profile.php
@@ -13,7 +13,7 @@ return new class extends Migration
     {
         Schema::create('offered_services', function (Blueprint $table) {
             $table->id();
-            $table->integer('user_id');
+            $table->unsignedBigInteger('user_id');
             $table->foreign('user_id')->references('id')->on('users')->onDelete('cascade');
             $table->string('description');
             $table->float('price');
@@ -21,9 +21,9 @@ return new class extends Migration
 
         Schema::create('avaliations', function (Blueprint $table) {
             $table->id();
-            $table->integer('from');
+            $table->unsignedBigInteger('from');
             $table->foreign('from')->references('id')->on('users')->onDelete('no action');
-            $table->integer('to');
+            $table->unsignedBigInteger('to');
             $table->foreign('to')->references('id')->on('users')->onDelete('no action');
             $table->integer('stars');
             $table->string('message',3000);
